list of possibly useful file formats for AsspDataObj corresponding to the first element of the fileInfo attribute

list of possibly useful file formats for AsspDataObj corresponding to the first element of the fileInfo attribute

Usage

AsspFileFormats

Format

Code Name	code number	description

RAW	1	headerless or unsupported format

ASP_A	2	ASP with ASCII data

ASP_B	3	ASP with binary data

XASSP	4	xassp ASCII

IPDS_M	5	labels in IPdS MIX' format \cr IPDS_S \tab 6\tab labels in IPdS SAMPA' format

AIFF	7	Apple Audio Interchange File Format

AIFC	8	AIFF extended for compressed data

CSL	9	Kay Elemetrics Computerized Speech Lab

CSRE	10	Computerized Speech Research Environment

ESPS	11	Entropic Signal Processing System

ILS	12

KTH	13	Kungliga Tekniska Hoegskolan Stockholm

SWELL	13	commercial version of KTH

SNACK	13	as Tcl extension

SFS	14	University College London Speech Filing System

SND	15	NeXT version of SND' format \cr AU \tab 15\tab Sun version of SND' format

NIST	16	National Institute of Standards and Technology

SPHERE	16	SPeech HEader REsources

PRAAT_S	17	UvA praat 'short' text file

PRAAT_L	18	UvA praat 'long' text file

PRAAT_B	19	UvA praat binary file

SSFF	20	Simple Signal File Format

WAVE	21	IBM/Microsoft RIFF-WAVE

WAVE_X	22	RIFF-WAVE extended format (Revision 3)

XLABEL	24	ESPS xlabel

YORK	25	University of York (Klatt'80 parameters)

UWM	26	University of Wisconsin at Madison (microbeam data) )

See also

Author

Lasse Bombien