Compute pitch and periodicity using the CREPE pitch tracker

The CREPE (Kim et al. 2018) applies a deep convolutional neural network directly on the time-domain waveform to find the fundamental frequency in a speech signal. Two versions of the models have been trained, one smaller yielding quicker results, and the full model which can be considerably more computationally intensive to apply.

Usage

crepe(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  windowSize = 15,
  minF = 70,
  maxF = 200,
  voicing.threshold = 0.21,
  silence.threshold = -60,
  model = c("tiny", "full"),
  explicitExt = "crp",
  outputDirectory = NULL,
  toFile = TRUE,
  conda.env = NULL
)

Arguments

listOfFiles

A vector of file paths to wav files.

beginTime

(Not implemented) The start time of the section of the sound file that should be processed.

endTime

(Not implemented) The end time of the section of the sound file that should be processed.

windowShift

The measurement interval (frame duration), in seconds.

windowSize

the analysis window length (in ms).

minF

Candidate f0 frequencies below this frequency will not be considered.

maxF

Candidates above this frequency will be ignored.

voicing.threshold

Voice/unvoiced threshold. Default is 0.21.

silence.threshold

Frames that do not contain amplitudes above this threshold (relative to the global maximum amplitude), are probably silent.

model

Use a fast ("tiny") model, or a more complete ("full") model to find pitch. The more complete model will take approximately 9-11 times longer to process the file.

explicitExt

the file extension that should be used.

outputDirectory

set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.

toFile

write the output to a file? The file will be written in outputDirectory, if defined, or in the same directory as the soundfile.

conda.env

The name of the conda environment in which Python and its required packages are stored. Please make sure that you know what you are doing if you change this. Defaults to NULL, which means that the default enviroment or the environment set in the RETICULATE_PYTHON environment variable will be used.

Value

An SSFF track object containing two tracks (f0 and periodicity) that are either returned (toFile == FALSE) or stored on disk.

Details

T

References

Kim JW, Salamon J, Li P, Bello JP (2018). “Crepe: A Convolutional Representation for Pitch Estimation.” 2018 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP), 00, 161--165. doi:10.1109/icassp.2018.8461329 .