Compute f0 using the Harvest algorithm

The Harvest algorithm (Morise 2017) was developed for the WORLD vocoder (MORISE et al. 2016) and aims to obtain a reliable F0 contour and reduce erroneously identified voice frames.

Usage

harvest(
  listOfFiles,
  beginTime = 0,
  endTime = 0,
  windowShift = 5,
  minF = 70,
  maxF = 200,
  explicitExt = "hf0",
  outputDirectory = NULL,
  toFile = TRUE
)

Arguments

listOfFiles

A vector of file paths to wav files.

beginTime

The start time of the section of the sound file that should be processed.

endTime

The end time of the section of the sound file that should be processed.

windowShift

The measurement interval (frame duration), in seconds.

minF

Candidate f0 frequencies below this frequency will not be considered.

maxF

Candidates above this frequency will be ignored.

explicitExt

the file extension that should be used.

outputDirectory

set an explicit directory for where the signal file will be written. If not defined, the file will be written to the same directory as the sound file.

toFile

write the output to a file? The file will be written in outputDirectory, if defined, or in the same directory as the soundfile.

Value

An SSFF track object containing two tracks (f0 and corr) that are either returned (toFile == FALSE) or stored on disk.

Details

The algorithm consists of two steps. In the first step, the algorithm uses fundamental component extraction by many band-pass filters with different center frequencies and obtains the basic f0 candidates from filtered signals. In the second step, basic f0 candidates are refined and scored by using the instantaneous frequency, and then several f0 candidates in each frame are estimated.

References

Morise M (2017). “Harvest: A High-Performance Fundamental Frequency Estimator from Speech Signals.” Interspeech 2017, 2321--2325. doi:10.21437/interspeech.2017-68 .

MORISE M, YOKOMORI F, OZAWA K (2016). “WORLD: A Vocoder-Based High-Quality Speech Synthesis System for Real-Time Applications.” IEICE Transactions on Information and Systems, E99.D(7), 1877--1884. ISSN 0916-8532, doi:10.1587/transinf.2015edp7457 .