Compute the Acoustic Voice Quality Index (AVQI)

This function computes the Acoustic Voice Quality Index (AVQI) from a set of continuous speech and sustained vowel samples. Praat is used to compute the AVQI and the result is therefore identical to the code published in (Latoszek et al. 2019) . The user may provide multiple continuous speech and sustained vowel samples for the speaker, and these will then be concatenates together before computing the AVQI value for the speaker and in that recording session.

Usage

praat_avqi(
  svDF,
  csDF,
  min.sv = 1000,
  speaker.name = NULL,
  speaker.ID = speaker.name,
  speaker.dob = NULL,
  session.datetime = NULL,
  pdf.path = NULL,
  simple.output = FALSE,
  overwrite.pdfs = FALSE,
  praat_path = NULL
)

Arguments

svDF

A data.frame containing columns "absolute_file_path","start", and "end". Each row should contain the full path to a sound file, and "start" and "end" points of a sustained vowel in that sound file. The multiple sustained vowels will be concatenated together to compute the AVQI, and it is therefore important that the user that the start and end points are indeed inside of the sustained vowel as to not influence the results by introducing unvoiced frames.

csDF

A data.frame containing columns "absolute_file_path","start", and "end". Each row should contain the full path to a sound file, and "start" and "end" points of a portion of continuous speech in that sound file. Multiple portions of speech will be concatenated together to compute the AVQI, and it is therefore important that the user that the start and end points are indeed inside of the portion of produced speech as to not influence the results by introducing unvoiced frames.

min.sv

The minimal continuous vowel duration required to make accurate measurements (in milliseconds). If the total duration of sustained vowels in the svDF tibble is smaller than this, the function will produce an error and quit processing. Defaults to 1000 ms (1 second) and should rarely be shorter than that.

speaker.name

The name of the speaker. Only used for in produced PDF output files.

speaker.ID

This will be used to identify the output returned list and in PDF output, and could therefore be either the ID of a speaker or of a Speaker + session compilation.

speaker.dob

The date of birth of the speaker. If provided, the PDF will be marked with this information.

session.datetime

The date and time when the recording was performed (as a string). If provided, the PDF will be marked with this information.

pdf.path

This is where PDF output files will be stored. If not provided, no PDF files will be produced.

simple.output

The AVQI Praat function can produce either a full page report of the voice profile, or a much more condensed version filling just a portion of the page. If simple.output=TRUE, the simplified version will be produced.

overwrite.pdfs

Should existing PDF files be overwritten in the PDF output directory? Defaults to a safe behavior where older PDFs are not overwritten.

praat_path

An explicit path to the Praat binary.

Value

A list with the following fields:

ID

The speaker / speaker + session identifier of the output

CPPS

Smoothed Cepstral Peak Prominence value

HNR

An Harmonic-to-noise estimate

Shim_local

A (local) Shimmer measurement (in %)

Shim_local_DB

A (local) Shimmer measurement, in decibels

LTAS_Slope

The slope of the Long Time Average Spectrum (in dB)

LTAS_Tilt

The Long Time Average Spectrum tilt (in dB)

AVQI

Acoustic Voice Quality Index summarizing the measures above

Details

If the user provides an pdf.path, a PDF of the AVQI analysis as implemented in the Praat script of (Latoszek et al. 2019) will be placed there. The name of the PDF will be speaker.ID_session.datetime, or "NA" if not provided. The default behavior is to not overwrite existing PDF files, as this could result in loss of data. The user should make sure that an appropriate speaker.ID and session.datetime values are provided to mark the output appropriately.

References

Latoszek BBv, Ulozaitė‐Stanienė N, Petrauskas T, Uloza V, Maryn Y (2019). “Diagnostic Accuracy of Dysphonia Classification of DSI and AVQI.” The Laryngoscope, 129(3), 692--698. ISSN 0023-852X, doi:10.1002/lary.27350 , http://www.ncbi.nlm.nih.gov/pubmed/30203473.