Compute the Dysphonia Severity Index

This function computes the Dysphonia Severity Index (DSI) (Wuyts et al. 2000-06) as implemented into Praat by Maryn et al. (2017) . The user is asked to supply data.frames indicating the "absolute_file_path" as well as "start", and "end" times for parts of sound files that together form the basis for the DSI computation for a speaker.

Usage

praat_dsi(
  softDF,
  highpitchDF,
  maxprolongedDF,
  stableDF = NULL,
  use.calibration = FALSE,
  db.calibration = 10,
  speaker.name = NULL,
  speaker.ID = NULL,
  speaker.dob = NULL,
  session.datetime = NULL,
  pdf.path = NULL,
  overwrite.pdfs = FALSE,
  praat_path = NULL
)

Arguments

softDF

A data.frame containing the file path and start and end times of samples indicating the softest voice production of the speaker.

highpitchDF

A data.frame containing the file path and start and end times of samples from which the highest f0 of the speaker can be computed.

maxprolongedDF

A data.frame containing the file path and start and end times of samples indicating maximally prolonged vowels of the speaker.

stableDF

An optional data.frame containing the file path and start and end times of samples indicating a stable sustained vowel. If not provided, the maxprolongedDF will be used instead.

use.calibration

Should a calibration factor be added to measured intensities before computing DSI?

db.calibration

The number of dB:s to be added to measured intensity values before computing DSI.

speaker.name

An optional name of the speaker. Only used when PDF output is produced.

speaker.ID

An ID of the speaker. This will be included in the list output.

speaker.dob

An optional date of birth of the speaker. Only used when PDF output is produced.

session.datetime

A string indicating the session date (and time). Only used when PDF output is produced.

pdf.path

A path where a PDF report file will be produced.

overwrite.pdfs

Should existing PDF files be overwritten in the PDF output directory? Defaults to a safe behavior where older PDFs are not overwritten.

praat_path

An explicit path to the Praat binary.

Value

A list with the following fields:

ID

The speaker speaker + session identifier of the output

Maximum.phonation.time

The speakers maximum phonation time.

Softest.intensity.of.voiced.speech

The intensity of the participants softest voice (in dB)

Maximum.fundamental.frequency

The maximum f0 frequency (in Hz)

Jitter.ppq5

The five-point Period Pertubation Quotient measurement (in %)

Dysphonia.Severity.Index

The resulting Dysphonia Severity Index

A meta-analysis in which 1330 healthy participants (aged 17.3−94) were included indicated an average DSI of 3.05 in healthy speakers, with a conﬁdence level between 2.13 and 3.98 (Sobol and Sielska-Badurek 2020) .

Details

The user has to indicate at least one sample in which the participant speaks as softly as possible (softDF), at least one sample from which the maximum f$_0$ could be deduced (highpitchDF), at least one sample where a single vowel is maximally prolonged (maxprolongedDF). The user can also provide a sample of a maximally stable vowel (stableDF), but if such a sample is not provided then the (maxprolongedDF) sample will be reused instead for the computation of the DSI sub-component Jitter. The user may submit multiple sound samples for all these sets of acoustic inputs, and the sounds files will then be combined before the DSI sub components are computed. The largest Maximum performance time will be used.

References

Maryn Y, Morsomme D, Bodt MD (2017). “Measuring the Dysphonia Severity Index (DSI) in the Program Praat.” Journal of Voice, 31(5), 644.e29--644.e40. ISSN 0892-1997, doi:10.1016/j.jvoice.2017.01.002 , http://www.ncbi.nlm.nih.gov/pubmed/28187924.

Sobol M, Sielska-Badurek EM (2020). “The Dysphonia Severity Index (DSI)—Normative Values. Systematic Review and Meta-Analysis.” Journal of Voice. ISSN 0892-1997, doi:10.1016/j.jvoice.2020.04.010 , http://www.ncbi.nlm.nih.gov/pubmed/32381275.

Wuyts FL, Bodt MSD, Molenberghs G, Remacle M, Heylen L, Millet B, Lierde KV, Raes J, Heyning PHVd (2000-06). “The Dysphonia Severity Index: an objective measure of vocal quality based on a multiparameter approach.” Journal of Speech, Language and Hearing Research, 43(3), 796 -- 809. doi:10.1044/jslhr.4303.796 , https://pubs.asha.org/doi/abs/10.1044/jslhr.4303.796.