Call the 'praat_sauce' analysis bundle to generate SSFF tracks

This function applies the praat_sauce bundle of Praat scripts on a single file and puts the resulting data as tracks in an SSFF file. By default, all analyses are applied to a windowed portion of the signal every 5ms.

Usage

praat_sauce(
  listOfFiles,
  beginTime = NULL,
  endTime = NULL,
  windowShift = 5,
  windowSize = 25,
  minF = 50,
  maxF = 300,
  formantTracking = TRUE,
  numFormants = 5,
  maxFormantHz = 5000,
  nominalF1 = 500,
  nominalF2 = 1500,
  nominalF3 = 2500,
  preEmphFrom = 50,
  useBandwidthFormula = FALSE,
  channel = 1,
  resample_to_16k = TRUE,
  toFile = TRUE,
  explicitExt = "psa",
  outputDirectory = NULL,
  verbose = FALSE,
  praat_path = NULL
)

Arguments

listOfFiles

A vector of file names

beginTime

The time in the sound file where analysis should start.

endTime

The last time point to be included in the analysed sample. If zero (0), the sound file will be included until the end.

windowShift

The time shift to the next analysis window. Defaults to every 5ms.

windowSize

The analysis window length (in ms).

minF

The minimal f0 to search for.

maxF

The maximum f0 to search for.

formantTracking

Boolean; Should the formant tracking abilities of Praat be used? Defaults to TRUE. If disabled, the raw window-by-window formant values will be used.

numFormants

The number of formants to be found within the frequency space.

maxFormantHz

The cutoff frequency used when finding the numFormants formants.

nominalF1

The nominal F1 used in formant tracking.

nominalF2

The nominal F2 used in formant tracking.

nominalF3

The nominal F3 used in formant tracking.

preEmphFrom

The frequency from which pre-emphasis will be applied.

useBandwidthFormula

Should the bandwidth calculation metod of Hawks and Miller (1995) be used, instead of Praat's internal algorithm. Defaults to TRUE (use Hawks & Miller's method).

channel

Which channel to use analyse. Defaults to the first channel.

resample_to_16k

Resample the signal to 16000 Hz before processing? Defaults to TRUE.

toFile

Should the SSFF track file be written to disk? Defaults to true, which by default means that it will be written as a file with a explicitExt extension next to the sound file.

explicitExt

The default file extension to use for the SSFF file.

outputDirectory

A path to an alternative output directory.

verbose

Verbose output. Not currently used.

praat_path=NULL

Value

This function builds an SSFF track object and writes it to disk, or returns it (toFile==FALSE). The track object will contain tracks with these fields:

f0

A track of the fundamental frequency computed by Praat from the signal as part of the PraatSauce analysis.

fm

Computed formant frequency measures, optionally tracked, one column per formant.

bw

Computed formant bandwiths, one column per formant.

H

The amplitudes of the first three harmonics, computed without accounting for the influence of adjecent formants (uncorrected).

Hc

The amplitudes of the first three harmonics, in which the influence of adjecent formants has been accounted for (corrected).

A

The amplitudes of the three harmonics that are closest to the center frequency of the first three formants, computed without accounting for the influence of the adjecent formants (uncorrected).

Ac

The amplitudes of the three harmonics that are closest to the center frequency of the first three formants, in which the influence of the adjecent formants has been accounted for (corrected).

HH

The differences between the (uncorrected) amplitudes of the first and second (column 1) and second and fourth (column 2).

HHc

The differences between the amplitudes of the first and second (column 1) and second and fourth (column 2), computed with correction with regards to neighbouring formants.

HA

The differences between the (uncorrected) amplitudes of the harmonics closest to the center frequencies of the first and second (column 1) and second and fourth (column 2) formants.

HAc

The differences between the amplitudes of the harmonics closest to the center frequencies of the first and second (column 1) and second and fourth (column 2) formants, corrected for the influence of the adjecent formants themselves.

cpp

A track of containing the smoothed Cepstral Peak Prominence (Fraile and Godino-Llorente 2014; Hillenbrand et al. 1994) across the acoustic signal.

hnr

Harmonic-to-noise ratios as defined by (Krom 1993) for frequencies 0-500 Hz, 0-1500 Hz, 0-2500 Hz, and 0-3500 Hz, respectively. (four columns)

References

Fraile R, Godino-Llorente JI (2014). “Cepstral peak prominence: A comprehensive analysis.” Biomedical Signal Processing and Control, 14, 42--54. ISSN 1746-8094, doi:10.1016/j.bspc.2014.07.001 .

Hawks JW, Miller JD (1995). “A formant bandwidth estimation procedure for vowel synthesis.” The Journal of the Acoustical Society of America, 97(2), 1343--1344. ISSN 0001-4966, doi:10.1121/1.412986 , http://www.ncbi.nlm.nih.gov/pubmed/7876453.

Hillenbrand J, Cleveland RA, Erickson RL (1994). “Acoustic Correlates of Breathy Vocal Quality.” Journal of Speech, Language, and Hearing Research, 37(4), 769--778. ISSN 1092-4388, doi:10.1044/jshr.3704.769 , http://www.ncbi.nlm.nih.gov/pubmed/7967562.

Krom Gd (1993). “A Cepstrum-Based Technique for Determining a Harmonics-to-Noise Ratio in Speech Signals.” Journal of Speech, Language, and Hearing Research, 36(2), 254--266. ISSN 1092-4388, doi:10.1044/jshr.3602.254 , http://www.ncbi.nlm.nih.gov/pubmed/8487518.

Examples

if (FALSE) {

}