Computes 18 measures of (vocal) tremor

This function calls Praat and applies the "tremor 3.05" package (Brückl 2021) to compute measures of voice tremor (Brückl et al. 2017) on a sustained vowel. The user may compute the measures for the entire sample or identify a portion of the sound file where the sustained is produced (beginTime and endTime arguments). The user may additionally specify an offset (selectionOffset) from the beginTime where measurements should start, as well as a (maximum) selection length. Using these arguments the user may mark where all sustained vowels are produced in a database, and apply this function across only a 2 second sample from 1 second into the vowel, and thus avoid vowel initial effects in phonation affecting measurements.

Usage

praat_voice_tremor(
  listOfFiles,
  beginTime = NULL,
  endTime = NULL,
  selectionOffset = NULL,
  selectionLength = 3,
  windowShape = "Gaussian1",
  relativeWidth = 1,
  minF = 75,
  maxF = 600,
  windowShift = 5,
  max_period_factor = 1.3,
  max_ampl_factor = 1.6,
  silence_threshold = 0.03,
  voicing_threshold = 0.45,
  octave_cost = 0.01,
  octave_jump_cost = 0.35,
  voiced_unvoiced_cost = 0.14,
  envelope.amplitude = TRUE,
  min.tremor.hz = 1.5,
  max.tremor.hz = 15,
  contour.magnitude.threshold = 0.01,
  tremor.cyclicality.threshold = 0.15,
  freq.tremor.octave.cost = 0.01,
  ampl.tremor.octave.cost = 0.01,
  na.zero = FALSE,
  praat_path = NULL
)

Arguments

listOfFiles

The full path of the sound file.

beginTime

The time point (in s) in the sound file where the sustained vowel starts. If NULL, the start of the sound file will also be viewed as the start of the sustained vowel production.

endTime

The time point (in s) in the sound file where the sustained vowel ends. If NULL, everyting up until the end of the sound file will be considered part of the sustained vowel.

selectionOffset

An optional offset to be added to the time of the sustained vowel production when determining where the start of the extracted portion of the vowel.

selectionLength

An optional (maximal) length of the selection.

windowShape

The window shape used for extracting the vowel. May be one of "rectangular", "triangular", "parabolic", "Hanning", "Hamming", "Gaussian1", "Gaussian2", "Gaussian3", "Gaussian4", "Gaussian5", "Kaiser1", and "Kaiser2".

relativeWidth

The relative width of the window used for extracting the vowel portion.

minF

The minimum pitch (f~0~) to be considered.

maxF

The maximum pitch (f~0~) to be considered.

windowShift

The number of milliseconds between analysis window centers.

max_period_factor

The larges possible differences between consecutive intervals that will be used in computing jitter. Please consult the Praat manual for further information.

max_ampl_factor

The larges possible differences between consecutive intervals that will be used in computing schimmer Please consult the Praat manual for further information.

silence_threshold

The silence threshold. Please consult the Praat manual for further information.

voicing_threshold

The voicing threshold. Please consult the Praat manual for further information.

octave_cost

The octave cost. Please consult the Praat manual for further information.

octave_jump_cost

The octave jump cost. Please consult the Praat manual for further information.

voiced_unvoiced_cost

The cost for voiced to unvoiced change detection. Please consult the Praat manual for further information.

na.zero

Should measurements of voice tremor that could not be determined be returned as a zero value, or as an NA (the default).

praat_path

An optional explicit path to the Praat binary. Not usually required.

Value

A list of voice tremor measurements :

FCoM

frequency contour magnitude

FTrC

(maximum) frequency tremor cyclicality

FMon

number of frequency modulations above thresholds

FTrF Hz

(strongest) frequency tremor frequency

FTrI %

frequency tremor intensity index

FTrP

frequency tremor power index

FTrCIP

frequency tremor cyclicality intensity product

FTrPS

frequency tremor product sum

FCoHNRdB

frequency contour harmonicity-to-noise ratio

ACoM

amplitude contour magnitude

ATrC

(maximum) amplitude tremor cyclicality

AMoN

number of amplitude modulations above thresholds

ATrF Hz

(strongest) amplitude tremor frequency

ATrI %

amplitude tremor intensity index

ATrP

amplitude tremor power index

ATrCIP

amplitude tremor cyclicality intensity product

ATrPS

amplitude tremor product sum

ACoHNRdB

amplitude contour harmonicity-to-noise ratio

References

Brückl MAE (2021). “Praat-script for the computation of 18 measures of (vocal) tremor.” doi:10.13140/RG.2.2.13850.57287 .

Brückl MAE, Ibragimova E, Bögelein S (2017). “Acoustic Tremor Measurement: Comparing Two Systems.” In Proceedings of the 10th International Workshop on Models and Analysis of Vocal Emissions for Biomedical Applications (MAVEBA), 19--22.