superassp: speech signal processing using various frameworks using a wrassp-like interface

This package bundles together routines that utilizes Praat, matlab or python to do the signal processing. The output of the original routines are then wrapped into a common interface that outputs an SSFF file or returns a wrassp compatible AsspDataObj object.

See also

Author

Maintainer: Fredrik Nylén fredrik.nylen@umu.se (ORCID)

Authors:

Fredrik Nylén fredrik.nylen@umu.se (ORCID)